refactor: extract indexer helpers, remove raw prominence-component fields (ADR-0027)
Extract from PlanetSearchProfile into focused, unit-testable units:
- AccountingBulkListener — the bulk-outcome counting/retry listener (was an
  anonymous inner class)
- IndexingStats — the LongAdder accounting counters + two-bucket failure
  thresholds
- OsmTagUtils — pure OSM-tag helpers (buildAltNames, classifyFeatureClass,
  parseFirstNumber) + OsmTagCache

Remove the five index:false raw prominence-component fields
(prom_base/prom_qrank_norm/prom_meta/ele_norm/qrank_raw) from the ES mapping,
PointDocument, and setProminence (ADR-0027). Production scoring reads only the
composite prominence; elevation still feeds it via ProminenceCalculator.

Code-review follow-ups:
- thread one OsmTagCache per feature through convertTagsToDocument +
  setIconColorCategory (was built twice per icon feature)
- AccountingBulkListener takes the IndexingStats holder instead of six
  positional LongAdders at the production call site
- fix stale comments on setProminence / ProminenceCalculator.Result

mvn test: 102 tests, 0 failures.
