+++ src/main/java/il/org/osm/israelhiking/PlanetSearchProfile.java
@@ -3,15 +3,12 @@ package il.org.osm.israelhiking;
 import static com.onthegomap.planetiler.reader.osm.OsmElement.Type.RELATION;
 import static com.onthegomap.planetiler.reader.osm.OsmElement.Type.WAY;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
-import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.atomic.LongAdder;
 import java.util.logging.Logger;
 import java.util.stream.Collectors;
 
@@ -31,22 +28,9 @@ import com.onthegomap.planetiler.reader.WithTags;
 import com.onthegomap.planetiler.reader.osm.OsmElement;
 import com.onthegomap.planetiler.reader.osm.OsmRelationInfo;
 
-import java.io.IOException;
-import java.net.ConnectException;
-import java.net.SocketTimeoutException;
-import java.util.concurrent.ThreadLocalRandom;
-
-import org.apache.http.conn.ConnectTimeoutException;
-import org.elasticsearch.client.ResponseException;
-
 import co.elastic.clients.elasticsearch.ElasticsearchClient;
 import co.elastic.clients.elasticsearch._helpers.bulk.BulkIngester;
-import co.elastic.clients.elasticsearch._helpers.bulk.BulkListener;
-import co.elastic.clients.elasticsearch._types.ElasticsearchException;
-import co.elastic.clients.elasticsearch.core.BulkRequest;
-import co.elastic.clients.elasticsearch.core.BulkResponse;
 import co.elastic.clients.elasticsearch.core.bulk.BulkOperation;
-import co.elastic.clients.elasticsearch.core.bulk.BulkResponseItem;
 
 public class PlanetSearchProfile implements Profile {
   private static final Logger LOGGER = Logger.getLogger(PlanetSearchProfile.class.getName());
@@ -56,52 +40,24 @@ public class PlanetSearchProfile implements Profile {
   private final String pointsIndexName;
   private final String bboxIndexName;
   private final String[] supportedLanguages;
-  // QRank lookup (Wikimedia pageviews by wikidata id) for the prominence signal.
-  // Empty index when no QRank file is provided (local builds) — lookups return 0.
+  // QRank lookup (Wikimedia pageviews by wikidata id) for the prominence signal. Empty index when
+  // no QRank file is provided (local builds), so lookups return 0.
   private final QRankIndex qrankIndex;
 
-  // Bulk indexer: thread-safe, buffers operations and flushes them in batches.
-  // Planetiler emits features from multiple worker threads, so BulkIngester
-  // (which handles concurrent add() internally) is the right tool rather than
-  // manually batching BulkRequests with our own locking.
+  // Bulk indexer: thread-safe, buffers operations and flushes them in batches. Planetiler emits
+  // features from multiple worker threads, so BulkIngester (concurrent add() internally) fits
+  // better than hand-locked BulkRequests.
   private final BulkIngester<Void> bulkIngester;
-  // Counters so indexing is observable instead of failing silently.
-  private final LongAdder indexedCount = new LongAdder();
-  // Failures are split by destination index so we can fail the build on the ones
-  // that cause missing SEARCH results (points) while only warning on bbox
-  // geo_shape rejects (a few degenerate OSM relation geometries ES can't parse;
-  // they don't affect name search). failedCount stays the SUM so the lossless
-  // invariant emitted == indexed + failed still holds.
-  private final LongAdder failedPointsCount = new LongAdder();
-  private final LongAdder failedBboxCount = new LongAdder();
-  // Step D — indexer resilience: a SEPARATE bucket for points ops charged only
-  // AFTER the bounded retry/backoff in the whole-batch (Throwable) path is
-  // exhausted. These are TRANSIENT charges (client-side connection/socket
-  // timeouts, 429/5xx) — likely phantom (ES may have committed them after the
-  // client gave up). They are tolerated far more generously than genuine per-item
-  // data failures, and the post-build reconcile gate is the real backstop. Kept
-  // distinct from failedPointsCount so a real mass mapping break (per-item 4xx)
-  // still trips the strict guard. transientBbox is the bbox analogue (warn-only).
-  private final LongAdder transientPointsCharges = new LongAdder();
-  private final LongAdder transientBboxCharges = new LongAdder();
-  // Emitted/attempted docs: incremented once per bulkIngester.add(...). Lets us
-  // assert the lossless invariant emitted == indexed + failed after flush() so a
-  // doc that was emitted but never made it into any bulk batch is detectable.
-  private final LongAdder emittedCount = new LongAdder();
-  // Step D — emitted POINTS only (excludes bbox). The reconcile gate and the
-  // points failure thresholds must be denominated in points, not points+bbox.
-  private final LongAdder emittedPointsCount = new LongAdder();
+  // Indexing accounting counters + failure thresholds, kept in their own holder (not the focus of
+  // this class). PlanetSearchProfile increments the emitted buckets on each emit; the listener
+  // shares the same LongAdders to record indexed/failed/transient outcomes from the bulk callbacks.
+  private final IndexingStats stats = new IndexingStats();
 
   public static final String POINTS_LAYER_NAME = "global_points";
 
   private static final Map<String, MinWayIdFinder> Singles = new ConcurrentHashMap<>();
   private static final Map<String, MinWayIdFinder> NamedHighways = new ConcurrentHashMap<>();
   private static final Map<String, MinWayIdFinder> Waterways = new ConcurrentHashMap<>();
-
-  public PlanetSearchProfile(PlanetilerConfig config, ElasticsearchClient esClient, String pointsIndexName,
-      String bboxIndexName, String[] supportedLnaguages) {
-    this(config, esClient, pointsIndexName, bboxIndexName, supportedLnaguages, QRankIndex.empty());
-  }
 
   public PlanetSearchProfile(PlanetilerConfig config, ElasticsearchClient esClient, String pointsIndexName,
       String bboxIndexName, String[] supportedLnaguages, QRankIndex qrankIndex) {
@@ -117,308 +73,16 @@ public class PlanetSearchProfile implements Profile {
         .maxOperations(5_000)
         .maxSize(5 * 1024 * 1024)
         .maxConcurrentRequests(4)
-        // Listener gives us per-batch visibility instead of swallowing errors.
-        // Extracted into a named class (AccountingBulkListener) so the counting
-        // logic is directly unit-testable; behavior is unchanged.
-        .listener(new AccountingBulkListener(esClient, indexedCount, failedPointsCount, failedBboxCount,
-            transientPointsCharges, transientBboxCharges, pointsIndexName)));
-  }
-
-  /**
-   * BulkListener that surfaces and counts every bulk outcome instead of
-   * swallowing errors. Extracted from an anonymous inner class so the per-item
-   * classification (indexed vs failed) and the whole-batch failure path are
-   * unit-testable by driving {@code afterBulk} directly. The classification
-   * logic is unchanged from the original anonymous listener.
-   */
-  static final class AccountingBulkListener implements BulkListener<Void> {
-    // Step D — retry/backoff tuning for the whole-batch transient path.
-    static final int MAX_RETRY_ATTEMPTS = 5;          // total resubmits (1s,2s,4s,8s,16s)
-    static final long BASE_BACKOFF_MILLIS = 1_000L;   // first backoff
-    static final long MAX_BACKOFF_MILLIS = 16_000L;   // cap per attempt
-
-    /** Injectable sleep so the backoff is testable without real delays. */
-    @FunctionalInterface
-    interface Sleeper {
-      void sleep(long millis) throws InterruptedException;
-    }
-
-    private final ElasticsearchClient esClient;
-    private final LongAdder indexedCount;
-    private final LongAdder failedPointsCount;
-    private final LongAdder failedBboxCount;
-    private final LongAdder transientPointsCharges;
-    private final LongAdder transientBboxCharges;
-    private final String pointsIndexName;
-    private final Sleeper sleeper;
-
-    AccountingBulkListener(ElasticsearchClient esClient, LongAdder indexedCount, LongAdder failedPointsCount,
-        LongAdder failedBboxCount, LongAdder transientPointsCharges, LongAdder transientBboxCharges,
-        String pointsIndexName) {
-      this(esClient, indexedCount, failedPointsCount, failedBboxCount, transientPointsCharges,
-          transientBboxCharges, pointsIndexName, Thread::sleep);
-    }
-
-    // Test seam: lets unit tests inject a no-op/recording Sleeper.
-    AccountingBulkListener(ElasticsearchClient esClient, LongAdder indexedCount, LongAdder failedPointsCount,
-        LongAdder failedBboxCount, LongAdder transientPointsCharges, LongAdder transientBboxCharges,
-        String pointsIndexName, Sleeper sleeper) {
-      this.esClient = esClient;
-      this.indexedCount = indexedCount;
-      this.failedPointsCount = failedPointsCount;
-      this.failedBboxCount = failedBboxCount;
-      this.transientPointsCharges = transientPointsCharges;
-      this.transientBboxCharges = transientBboxCharges;
-      this.pointsIndexName = pointsIndexName;
-      this.sleeper = sleeper;
-    }
-
-    private boolean isPoints(String index) {
-      return pointsIndexName != null && pointsIndexName.equals(index);
-    }
-
-    // A failed item destined for the points index means a missing SEARCH result
-    // (build-breaking); a failed bbox item is a geometry reject (warn-only). This
-    // bucket is for GENUINE per-item data failures (4xx mapping/parse).
-    private void recordFailure(String index) {
-      if (isPoints(index)) {
-        failedPointsCount.increment();
-      } else {
-        failedBboxCount.increment();
-      }
-    }
-
-    // The TRANSIENT bucket: whole-batch ops charged only after retries are
-    // exhausted. Tolerated generously and backstopped by the reconcile gate.
-    private void recordTransient(String index) {
-      if (isPoints(index)) {
-        transientPointsCharges.increment();
-      } else {
-        transientBboxCharges.increment();
-      }
-    }
-
-    @Override
-    public void beforeBulk(long executionId, BulkRequest request, List<Void> contexts) {
-      // no-op
-    }
-
-    @Override
-    public void afterBulk(long executionId, BulkRequest request, List<Void> contexts, BulkResponse response) {
-      if (response.errors()) {
-        response.items().forEach(item -> {
-          if (item.error() != null) {
-            recordFailure(item.index());
-            LOGGER.warning(() -> "Failed to index id=" + item.id() + " into " + item.index()
-                + ": " + item.error().reason());
-          } else {
-            indexedCount.increment();
-          }
-        });
-      } else {
-        indexedCount.add(response.items().size());
-      }
-    }
-
-    @Override
-    public void afterBulk(long executionId, BulkRequest request, List<Void> contexts, Throwable failure) {
-      // A whole-batch failure has no per-item breakdown. The 8.x BulkIngester does
-      // NOT auto-resubmit on a Throwable, so we classify and (for transient errors)
-      // retry the batch with exponential backoff + jitter before charging anything.
-      if (isRetryable(failure)) {
-        LOGGER.warning(() -> "Bulk request " + executionId + " failed transiently ("
-            + describe(failure) + "); retrying with backoff.");
-        resubmitWithBackoff(executionId, request.operations());
-      } else {
-        // Non-retryable whole-batch error (e.g. a 4xx request-level rejection): do
-        // NOT retry (would loop forever) — charge as genuine data failures so a real
-        // structural problem trips the strict guard.
-        LOGGER.severe(() -> "Bulk request " + executionId + " failed non-retryably ("
-            + describe(failure) + "); charging " + request.operations().size()
-            + " op(s) as genuine data failures.");
-        request.operations().forEach(op -> recordFailure(bulkOperationIndex(op)));
-      }
-    }
-
-    /**
-     * Resubmit a timed-out batch's operations via a fresh synchronous bulk call,
-     * with exponential backoff + jitter, up to {@link #MAX_RETRY_ATTEMPTS} times.
-     * After each attempt the response is reconciled per item: successes are counted
-     * as indexed, per-item 4xx errors are charged as genuine data failures (never
-     * retried), and only ops that hit a retryable condition (or a whole-request
-     * retryable throwable) carry over to the next attempt. Whatever remains after
-     * the final attempt is charged to the TRANSIENT bucket.
-     */
-    private void resubmitWithBackoff(long executionId, List<BulkOperation> operations) {
-      List<BulkOperation> pending = new ArrayList<>(operations);
-      for (int attempt = 1; attempt <= MAX_RETRY_ATTEMPTS && !pending.isEmpty(); attempt++) {
-        long backoff = backoffMillis(attempt);
-        try {
-          sleeper.sleep(backoff);
-        } catch (InterruptedException ie) {
-          Thread.currentThread().interrupt();
-          LOGGER.warning("Retry interrupted; charging remaining " + pending.size()
-              + " op(s) as transient.");
-          break;
-        }
-        final List<BulkOperation> batch = pending;
-        try {
-          BulkResponse response = esClient.bulk(b -> b.operations(batch));
-          pending = reconcileRetryResponse(batch, response);
-          if (pending.isEmpty()) {
-            final int n = batch.size();
-            final int a = attempt;
-            LOGGER.info(() -> "Bulk request " + executionId + " recovered on retry attempt "
-                + a + " (" + n + " op(s) re-applied).");
-            return;
-          }
-        } catch (Exception e) {
-          if (!isRetryable(e)) {
-            // The whole resubmit hit a non-retryable error — charge all pending as
-            // genuine data failures and stop (retrying can't help).
-            LOGGER.severe(() -> "Retry of bulk request " + executionId
-                + " hit a non-retryable error (" + describe(e) + "); charging "
-                + batch.size() + " op(s) as genuine data failures.");
-            batch.forEach(op -> recordFailure(bulkOperationIndex(op)));
-            return;
-          }
-          final int a = attempt;
-          LOGGER.warning(() -> "Retry attempt " + a + " for bulk request " + executionId
-              + " failed transiently (" + describe(e) + ").");
-          // keep `pending` as-is for the next attempt
-        }
-      }
-      // Retries exhausted (or interrupted): charge whatever is still pending to the
-      // TRANSIENT bucket — likely phantom (ES may have committed despite the timeout),
-      // and the reconcile gate is the real backstop.
-      if (!pending.isEmpty()) {
-        final int n = pending.size();
-        LOGGER.warning(() -> "Bulk request " + executionId + " exhausted retries; charging "
-            + n + " op(s) to the transient bucket.");
-        pending.forEach(op -> recordTransient(bulkOperationIndex(op)));
-      }
-    }
-
-    /**
-     * Classify a retry-attempt's response: count successes, charge per-item 4xx as
-     * genuine data failures, and return the ops that should be retried again
-     * (per-item retryable status: 429 or 5xx). An op with no matching response item
-     * is conservatively kept for retry.
-     */
-    private List<BulkOperation> reconcileRetryResponse(List<BulkOperation> batch, BulkResponse response) {
-      // Walk per OP (not per response item): a response shorter than the batch must NOT silently
-      // drop the surplus ops. Whether response.errors() is set or not, an op with no matching
-      // response item is kept for retry so it can never vanish from the emitted/indexed/failed
-      // accounting (the lossless invariant the reconcile gate relies on).
-      List<BulkOperation> retryAgain = new ArrayList<>();
-      List<BulkResponseItem> items = response.items();
-      for (int i = 0; i < batch.size(); i++) {
-        BulkResponseItem item = i < items.size() ? items.get(i) : null;
-        if (item == null) {
-          retryAgain.add(batch.get(i)); // no response for this op — retry it, don't drop it
-        } else if (item.error() == null) {
-          indexedCount.increment();
-        } else if (isRetryableStatus(item.status())) {
-          retryAgain.add(batch.get(i));
-        } else {
-          // Genuine per-item failure (e.g. 400 mapping/parse) — must NOT be retried.
-          recordFailure(item.index());
-          LOGGER.warning(() -> "Failed to index id=" + item.id() + " into " + item.index()
-              + ": " + item.error().reason());
-        }
-      }
-      return retryAgain;
-    }
-
-    // Exponential backoff with full jitter, capped. attempt is 1-based.
-    static long backoffMillis(int attempt) {
-      long exp = BASE_BACKOFF_MILLIS << (attempt - 1); // 1s,2s,4s,8s,16s,...
-      long capped = Math.min(exp, MAX_BACKOFF_MILLIS);
-      // Full jitter in [capped/2, capped] keeps it bounded but de-synchronised
-      // across the 4 concurrent bulk workers.
-      long half = capped / 2;
-      return half + ThreadLocalRandom.current().nextLong(half + 1);
-    }
-
-    /**
-     * Retryable = transient transport/availability problems that a resubmit can
-     * fix: connection/socket timeouts, connection refused, generic IO, and HTTP
-     * 429/502/503/504. Non-retryable = 4xx request/data errors (esp. 400 mapping
-     * /parse) which would loop forever.
-     */
-    static boolean isRetryable(Throwable t) {
-      for (Throwable c = t; c != null; c = c.getCause()) {
-        if (c instanceof SocketTimeoutException
-            || c instanceof ConnectTimeoutException
-            || c instanceof ConnectException) {
-          return true;
-        }
-        if (c instanceof ResponseException re) {
-          return isRetryableStatus(re.getResponse().getStatusLine().getStatusCode());
-        }
-        if (c instanceof ElasticsearchException ese) {
-          return isRetryableStatus(ese.status());
-        }
-        if (c instanceof IOException) {
-          // Generic IO (connection reset, broken pipe, premature EOF) — transient.
-          return true;
-        }
-        if (c.getCause() == c) {
-          break; // guard against self-referential cause chains
-        }
-      }
-      return false;
-    }
-
-    static boolean isRetryableStatus(int status) {
-      return status == 429 || status == 502 || status == 503 || status == 504;
-    }
-
-    private static String describe(Throwable t) {
-      String msg = t.getMessage();
-      return t.getClass().getSimpleName() + (msg == null ? "" : ": " + msg);
-    }
-
-    // Best-effort extraction of the destination index from a bulk operation
-    // (index/create/update/delete variants). Null when it can't be determined,
-    // which the failure/transient recorders treat as non-points (bbox) — the
-    // conservative choice for the warn-only bucket.
-    private static String bulkOperationIndex(BulkOperation op) {
-      if (op.isIndex()) {
-        return op.index().index();
-      } else if (op.isCreate()) {
-        return op.create().index();
-      } else if (op.isUpdate()) {
-        return op.update().index();
-      } else if (op.isDelete()) {
-        return op.delete().index();
-      }
-      return null;
-    }
+        // Listener gives per-batch visibility instead of swallowing errors. Extracted into a named
+        // class (AccountingBulkListener) so the counting logic is unit-testable. Pass the whole
+        // IndexingStats holder so the counters are wired by name, not a swap-prone positional list.
+        .listener(new AccountingBulkListener(esClient, stats, pointsIndexName)));
   }
 
   /*
-   * The processing happens in 3 steps:
-   * 1. On the first pass through the input file, store relevant information from
-   * applicable OSM route relations and ways with mtb:name tag.
-   * 2. On the second pass, emit points for relation and mtb:name ways. Emit a
-   * point by merging all the ways and using the first point of the merged
-   * linestring.
-   * 
-   * Step 1)
-   *
-   * Planetiler processes the .osm.pbf input file in two passes. The first pass
-   * stores node locations, and invokes
-   * preprocessOsmRelation for reach relation and stores information the profile
-   * needs during the second pass when we
-   * emit map feature for ways contained in that relation.
-   * 
-   * Step 2)
-   *
-   * On the second pass through the input .osm.pbf file, for each way in a
-   * relation that we stored data about, emit a
-   * point with attributes derived from the relation as well as for ways with
-   * mtb:name tag.
+   * Two-pass processing: pass 1 stores node locations and (via preprocessOsmRelation) the relation
+   * /mtb:name way info the profile needs later; pass 2 emits a point per relation and mtb:name way,
+   * merging the ways and using the first point of the merged linestring.
    */
 
   static private final void CoalesceIntoMap(Map<String, String> map, String language, String... strings) {
@@ -433,95 +97,28 @@ public class PlanetSearchProfile implements Profile {
   }
 
   /**
-   * Language-neutral OSM variant-name tag bases (no {@code :<lang>} suffix). These also have
-   * per-language forms ({@code alt_name:he}, {@code official_name:en}, ...). {@code old_name} is
-   * DELIBERATELY EXCLUDED — it is a documented Nominatim noise source (a stale former name can
-   * overtake the real one). See ADR-0011.
-   */
-  static private final String[] ALT_NAME_TAG_BASES = {
-      "alt_name", "official_name", "short_name", "loc_name", "int_name"
-  };
-
-  /**
-   * Populate {@code pointDocument.alt_names} (ADR-0011 item 1) from the OSM variant-name tags,
-   * keyed by language exactly like {@code name}: the language-suffixed tags
-   * ({@code alt_name:<lang>}, ...) go under {@code <lang>}, and the bare tags ({@code alt_name},
-   * ...) go under {@code "default"}.
+   * Populate pointDocument.alt_names from the OSM variant-name tags, keyed by language like name:
+   * suffixed tags (alt_name:lang, ...) under lang, bare tags under "default".
    *
-   * <p>CRITICAL difference from {@link #CoalesceIntoMap}: alt-name tags are frequently MULTI-VALUED
-   * with {@code ;} (e.g. {@code alt_name=A;B;C}). CoalesceIntoMap keeps only the FIRST tag and never
-   * splits, so it cannot be reused here. We split every value on {@code ;}, trim, drop empties, and
-   * de-duplicate (preserving order) into a {@code List<String>} so each variant is a SEPARATE value
-   * (ES indexes a string array as independent tokens — no {@code ;}-joining, which would fold the
-   * variants back into one analyzed string). The map (and each language entry) is created lazily,
-   * so a feature with no variant tags leaves {@code alt_names} null (omitted from JSON).
+   * Unlike CoalesceIntoMap, alt-name tags are frequently multi-valued with ";" (alt_name=A;B;C), so
+   * we split on ";", trim, drop empties and de-dup (order-preserving) into a List so each variant is
+   * a separate token. Built lazily, so a feature with no variant tags leaves alt_names null.
    */
   private void addAltNames(PointDocument pointDocument, WithTags feature) {
-    var altNames = buildAltNames(supportedLanguages, feature::getString);
+    var altNames = OsmTagUtils.buildAltNames(supportedLanguages, feature::getString);
     if (altNames != null) {
       pointDocument.alt_names = altNames;
     }
   }
 
-  /**
-   * PURE (no I/O, no {@code WithTags}) builder for the {@code alt_names} map, so the {@code ;}-split,
-   * trim, de-dup and old_name-exclusion logic is directly unit-testable (mirrors the
-   * {@code flooredProminence} extraction pattern). Returns {@code null} when no variant tags are
-   * present (so {@code addAltNames} leaves {@code alt_names} null and it is omitted from JSON).
-   *
-   * @param supportedLanguages the per-language keys to look up suffixed tags for
-   * @param tagLookup          maps an OSM tag key to its raw value (null if absent)
-   */
-  static Map<String, List<String>> buildAltNames(String[] supportedLanguages,
-      java.util.function.Function<String, String> tagLookup) {
-    Map<String, List<String>> altNames = null;
-    for (String language : supportedLanguages) {
-      var suffixedTags = Arrays.stream(ALT_NAME_TAG_BASES)
-          .map(base -> base + ":" + language)
-          .toArray(String[]::new);
-      var collected = collectVariants(tagLookup, suffixedTags);
-      if (collected != null) {
-        if (altNames == null) {
-          altNames = new HashMap<String, List<String>>();
-        }
-        altNames.put(language, collected);
-      }
-    }
-    // Bare (language-neutral) tags -> "default".
-    var collectedDefault = collectVariants(tagLookup, ALT_NAME_TAG_BASES);
-    if (collectedDefault != null) {
-      if (altNames == null) {
-        altNames = new HashMap<String, List<String>>();
-      }
-      altNames.put("default", collectedDefault);
-    }
-    return altNames;
-  }
-
-  /**
-   * Read each of {@code tags}, split every value on {@code ;}, trim, drop empties, de-duplicate
-   * (insertion order preserved). Returns {@code null} if nothing survives (so the caller can skip
-   * the key entirely). {@code old_name} is never in {@code tags}.
-   */
-  private static List<String> collectVariants(java.util.function.Function<String, String> tagLookup,
-      String[] tags) {
-    var variants = new java.util.LinkedHashSet<String>();
-    for (String tag : tags) {
-      var raw = tagLookup.apply(tag);
-      if (raw == null || raw.isEmpty()) {
-        continue;
-      }
-      for (String part : raw.split(";")) {
-        var trimmed = part.trim();
-        if (!trimmed.isEmpty()) {
-          variants.add(trimmed);
-        }
-      }
-    }
-    return variants.isEmpty() ? null : new ArrayList<>(variants);
-  }
-
   private void convertTagsToDocument(PointDocument pointDocument, WithTags feature) {
+    convertTagsToDocument(pointDocument, feature, new OsmTagUtils.OsmTagCache(feature::getString));
+  }
+
+  // Cache-accepting overload: paths that also call setIconColorCategory on the same feature build one
+  // OsmTagCache and share it here (used for feature_class) and there, so the tag reads happen once.
+  private void convertTagsToDocument(PointDocument pointDocument, WithTags feature,
+      OsmTagUtils.OsmTagCache tags) {
     for (String language : supportedLanguages) {
       CoalesceIntoMap(pointDocument.name, language, feature.getString("name:" + language));
       CoalesceIntoMap(pointDocument.description, language, feature.getString("description:" + language));
@@ -532,9 +129,8 @@ public class PlanetSearchProfile implements Profile {
     if (feature.hasTag("description")) {
       CoalesceIntoMap(pointDocument.description, "default", feature.getString("description"));
     }
-    // ADR-0011 item 1 — variant names into the SEPARATE alt_names field (after the name loop).
-    // convertTagsToDocument is the single chokepoint that also serves relations (the IBT case),
-    // so this one call covers the relation path too.
+    // Variant names into the separate alt_names field. convertTagsToDocument is the single
+    // chokepoint that also serves relations, so this one call covers the relation path too.
     addAltNames(pointDocument, feature);
     setDifficulty(pointDocument, feature);
     pointDocument.wikidata = feature.getString("wikidata");
@@ -543,205 +139,19 @@ public class PlanetSearchProfile implements Profile {
     pointDocument.website = feature.getString("website");
     setProminence(pointDocument, feature);
     setPopulation(pointDocument, feature);
-    setFeatureClass(pointDocument, feature);
+    pointDocument.feature_class = OsmTagUtils.classifyFeatureClass(tags);
   }
 
   /**
-   * Derive a coarse FEATURE TYPE ("peak", "lake", "waterfall", "city", "spring", ...) from the
-   * primary OSM type tags and store it on {@code pointDocument.feature_class}. This is the kind-of-
-   * thing signal the engine needs to rank "Mount Shasta" as a SUMMIT, not a homonym hotel — distinct
-   * from poiCategory ("Wikipedia"/"Other"), which never identifies the feature type.
-   *
-   * <p>Reads the same tags {@link #setProminence} already consults (no new tag access). Most-specific
-   * tag wins; leaves the field null when nothing recognised is present (omitted from JSON, so the
-   * ~22.8M docs without a known type carry no extra bytes). Normalises OSM values into a small,
-   * stable vocabulary so the query side can match on a known set rather than raw OSM strings.
-   */
-  void setFeatureClass(PointDocument pointDocument, WithTags feature) {
-    pointDocument.feature_class = classifyFeatureClass(feature::getString);
-  }
-
-  /**
-   * PURE (no {@code WithTags}) core of {@link #setFeatureClass}, so the OSM-tag -> feature_class
-   * switch is directly unit-testable from a literal tag map (mirrors {@code buildAltNames} /
-   * {@code flooredProminence}). Returns null when nothing recognised is present.
-   */
-  static String classifyFeatureClass(java.util.function.Function<String, String> tagLookup) {
-    String natural = tagLookup.apply("natural");
-    String waterway = tagLookup.apply("waterway");
-    String place = tagLookup.apply("place");
-    String historic = tagLookup.apply("historic");
-    String tourism = tagLookup.apply("tourism");
-    // Built/POI keys (ADR-0019): consulted only AFTER the outdoor keys above, so an object
-    // carrying both an outdoor tag and an incidental built tag (e.g. tourism=viewpoint that is
-    // also building=yes) keeps its outdoor class. Unnamed objects never reach this code — every
-    // emit path gates on name (processOtherSourceFeature, addNonIconFeaturesToElasricseach), so
-    // building=yes noise is already filtered upstream and we classify only named features.
-    String leisure = tagLookup.apply("leisure");
-    String amenity = tagLookup.apply("amenity");
-    String shop = tagLookup.apply("shop");
-    String office = tagLookup.apply("office");
-    String craft = tagLookup.apply("craft");
-    String healthcare = tagLookup.apply("healthcare");
-    String manMade = tagLookup.apply("man_made");
-    String building = tagLookup.apply("building");
-
-    String fc = null;
-    if (natural != null) {
-      switch (natural) {
-        case "peak":                  fc = "peak"; break;
-        case "volcano":               fc = "peak"; break;
-        case "hill":                  fc = "hill"; break;
-        case "ridge":                 fc = "ridge"; break;
-        case "saddle": case "gap":    fc = "saddle"; break;
-        case "cliff":                 fc = "cliff"; break;
-        case "rock": case "stone":    fc = "rock"; break;
-        case "water":                 fc = "lake"; break;   // refined below by water=*
-        case "spring": case "hot_spring": fc = "spring"; break;
-        case "glacier":               fc = "glacier"; break;
-        case "bay":                   fc = "bay"; break;
-        case "cape":                  fc = "cape"; break;
-        case "beach":                 fc = "beach"; break;
-        case "wood": case "forest":   fc = "forest"; break;
-        case "valley":                fc = "valley"; break;
-        // landforms scoring already groups (TERRAIN_RELIEF) but the indexer never emitted:
-        case "canyon": case "gorge":  fc = "canyon"; break;
-        case "mesa": case "plateau":  fc = "plateau"; break;
-        case "arch":                  fc = "arch"; break;
-        case "cave_entrance":         fc = "cave"; break;
-        case "wetland":               fc = "wetland"; break;
-        case "arete":                 fc = "ridge"; break;    // fold: ridge kin
-        case "crater":                fc = "peak"; break;     // fold: volcanic-summit kin
-        case "mountain_range":        fc = "peak"; break;     // fold: high-ground kin
-        default:                      fc = "natural"; break;
-      }
-      // refine natural=water by its sub-type (lake vs reservoir vs pond)
-      String water = tagLookup.apply("water");
-      if ("water".equals(natural) && water != null) {
-        switch (water) {
-          case "lake":      fc = "lake"; break;
-          case "reservoir": fc = "reservoir"; break;
-          case "pond":      fc = "pond"; break;
-          default:          fc = "lake"; break;
-        }
-      }
-    } else if (waterway != null) {
-      switch (waterway) {
-        case "waterfall":             fc = "waterfall"; break;
-        case "river":                 fc = "river"; break;
-        case "stream":                fc = "stream"; break;
-        case "canal":                 fc = "canal"; break;
-        case "rapids":                fc = "rapids"; break;
-        default:                      fc = "waterway"; break;
-      }
-    } else if (place != null) {
-      switch (place) {
-        case "city": case "town": case "village": case "hamlet":
-        case "suburb": case "neighbourhood":
-                                      fc = place; break;
-        case "island": case "islet": fc = "island"; break;
-        case "locality":             fc = "locality"; break;
-        default:                      fc = "place"; break;
-      }
-    } else if (historic != null) {
-      fc = "historic";
-    } else if (tourism != null) {
-      switch (tourism) {
-        case "viewpoint":             fc = "viewpoint"; break;
-        case "camp_site":             fc = "campsite"; break;
-        case "attraction":            fc = "attraction"; break;
-        // lodging (built): folded into one "lodging" class
-        case "hotel": case "motel": case "hostel": case "guest_house":
-        case "apartment": case "chalet": case "alpine_hut": case "wilderness_hut":
-                                      fc = "lodging"; break;
-        case "museum": case "gallery": fc = "museum"; break;
-        case "information":           fc = "tourism"; break;
-        default:                      fc = "tourism"; break;
-      }
-    } else if (leisure != null) {
-      // Recreation: park/nature_reserve lean OUTDOOR (semi-outdoor group), sports the rest.
-      switch (leisure) {
-        case "park": case "garden": case "nature_reserve":
-        case "playground": case "dog_park": case "common":
-                                      fc = "park"; break;
-        case "sports_centre": case "stadium": case "pitch": case "track":
-        case "fitness_centre": case "swimming_pool": case "golf_course":
-        case "ice_rink": case "horse_riding":
-                                      fc = "sports"; break;
-        default:                      fc = "leisure"; break;
-      }
-    } else if (amenity != null) {
-      switch (amenity) {
-        case "restaurant": case "cafe": case "fast_food": case "bar":
-        case "pub": case "food_court": case "biergarten": case "ice_cream":
-                                      fc = "food"; break;
-        case "place_of_worship": case "monastery":
-                                      fc = "religious"; break;
-        case "school": case "university": case "college": case "kindergarten":
-        case "library":
-                                      fc = "education"; break;
-        case "hospital": case "clinic": case "pharmacy": case "doctors":
-        case "dentist": case "veterinary":
-                                      fc = "medical"; break;
-        case "townhall": case "courthouse": case "police": case "fire_station":
-        case "post_office": case "embassy": case "prison":
-                                      fc = "government"; break;
-        case "fuel": case "charging_station":
-                                      fc = "fuel"; break;
-        case "parking": case "parking_space": case "bicycle_parking":
-        case "motorcycle_parking": case "taxi":
-                                      fc = "parking"; break;
-        case "bus_station": case "ferry_terminal":
-                                      fc = "transit"; break;
-        case "theatre": case "cinema": case "arts_centre":
-                                      fc = "museum"; break;
-        default:                      fc = "amenity"; break;
-      }
-    } else if (shop != null) {
-      fc = "shop";
-    } else if (office != null) {
-      fc = "office";
-    } else if (craft != null) {
-      fc = "office";
-    } else if (healthcare != null) {
-      fc = "medical";
-    } else if (manMade != null) {
-      switch (manMade) {
-        case "tower": case "lighthouse": case "bridge": case "obelisk":
-        case "water_tower": case "windmill": case "watermill": case "pier":
-                                      fc = "structure"; break;
-        default:                      fc = "man_made"; break;
-      }
-    } else if (building != null && !"no".equals(building) && !"none".equals(building)
-               && !"No".equals(building)) {
-      // Named building with no more-specific type tag above. building=yes is the dominant
-      // (~80%) value; a named one is a real search target (a named hall/landmark), so it gets
-      // the generic "building" class rather than null.
-      switch (building) {
-        case "church": case "chapel": case "cathedral": case "mosque":
-        case "synagogue": case "temple": case "shrine":
-                                      fc = "religious"; break;
-        case "hotel":                 fc = "lodging"; break;
-        case "hospital":              fc = "medical"; break;
-        case "school": case "university": case "college":
-                                      fc = "education"; break;
-        case "train_station":         fc = "transit"; break;
-        default:                      fc = "building"; break;
-      }
-    }
-    return fc;
-  }
-
-  /**
-   * Compute the composite prominence score from OSM tags + QRank and store it (plus raw components,
-   * for re-tuning without a reindex) on the document. Reads tags directly from the feature rather
-   * than poiCategory, because poiCategory is assigned later in some emit paths.
+   * Compute the composite prominence score from OSM tags + QRank and store it on the document. Only
+   * the composite is stored — ADR-0027 removed the raw-component fields. Reads tags directly, not
+   * poiCategory, which is assigned later in some emit paths.
    */
   private void setProminence(PointDocument pointDocument, WithTags feature) {
     long qrankRaw = qrankIndex.getByWikidata(pointDocument.wikidata);
 
     // Elevation feeds the peak prominence prior, read from the OSM ele tag (NaN when absent).
-    double ele = parseFirstNumber(feature.getString("ele"));
+    double ele = OsmTagUtils.parseFirstNumber(feature.getString("ele"));
 
     boolean hasImage = pointDocument.image != null || pointDocument.wikimedia_commons != null;
     boolean hasWebsite = pointDocument.website != null;
@@ -757,23 +167,15 @@ public class PlanetSearchProfile implements Profile {
         ele, hasImage, hasWebsite, hasWikidata, qrankRaw);
 
     pointDocument.prominence = r.prominence;
-    pointDocument.prom_base = r.base;
-    pointDocument.prom_qrank_norm = r.qrankNorm;
-    pointDocument.prom_meta = r.meta;
-    pointDocument.ele_norm = r.eleNorm;
-    pointDocument.qrank_raw = r.qrankRaw > 0 ? r.qrankRaw : null;
 
     setEnrichmentSignals(pointDocument, feature);
   }
 
   /**
-   * ADR-0014 — additive ranking signals beyond the prominence composite:
-   * <ul>
-   *   <li>{@code area_norm}: log-normalized polygon area, only when the feature carries polygon
-   *       geometry (a {@link SourceFeature} that {@code canBePolygon()}). Left null for points/lines
-   *       and for relations routed through {@code convertTagsToDocument} as {@link WithTags}.</li>
-   *   <li>{@code intermittent}: the OSM {@code intermittent=yes} flag (seasonal/dry water).</li>
-   * </ul>
+   * Additive ranking signals beyond the prominence composite:
+   *   - area_norm: log-normalized polygon area, only for a polygon-capable SourceFeature; null for
+   *     points/lines and for relations routed through convertTagsToDocument as WithTags.
+   *   - intermittent: the OSM intermittent=yes flag (seasonal/dry water).
    */
   private void setEnrichmentSignals(PointDocument pointDocument, WithTags feature) {
     // area_norm — geometry-dependent, so only when we actually hold a polygon-capable SourceFeature.
@@ -781,10 +183,9 @@ public class PlanetSearchProfile implements Profile {
       try {
         pointDocument.area_norm = normalizeArea(sf.areaMeters());
       } catch (Exception e) {
-        // Bad polygon geometry — leave area_norm null, never fail the build. Logged at FINE
-        // rather than WARNING because this fires in the high-volume point-enrichment path and a
-        // few unbuildable polygons among ~22.8M features is an expected data condition; FINE keeps
-        // it diagnosable (enable il.org.osm.israelhiking at FINE) without flooding the console.
+        // Bad polygon geometry — leave area_norm null, never fail the build. Logged at FINE (not
+        // WARNING) because a few unbuildable polygons is an expected data condition in this
+        // high-volume path; FINE keeps it diagnosable without flooding the console.
         LOGGER.fine(() -> "area_norm skipped for " + sourceFeatureToDocumentId(sf)
             + " (" + e.getClass().getSimpleName() + ": " + e.getMessage() + ")");
       }
@@ -796,7 +197,7 @@ public class PlanetSearchProfile implements Profile {
     }
   }
 
-  /** {@code log1p(areaM)/log1p(MAX_AREA_M2)} clamped to [0,1]; MAX_AREA_M2 = 1e11 (~100k km²). */
+  /** log1p(areaM)/log1p(MAX_AREA_M2) clamped to [0,1]; MAX_AREA_M2 = 1e11 (~100k km²). */
   private static float normalizeArea(double areaM) {
     if (Double.isNaN(areaM) || areaM <= 0) {
       return 0f;
@@ -811,7 +212,7 @@ public class PlanetSearchProfile implements Profile {
     if (place == null) {
       return;
     }
-    double parsed = parseFirstNumber(feature.getString("population"));
+    double parsed = OsmTagUtils.parseFirstNumber(feature.getString("population"));
     if (!Double.isNaN(parsed) && parsed > 0) {
       pointDocument.population = (int) Math.min(parsed, Integer.MAX_VALUE);
       return;
@@ -824,41 +225,6 @@ public class PlanetSearchProfile implements Profile {
       case "village": pointDocument.population = 2_000; break;
       case "hamlet":  pointDocument.population = 200; break;
       default:        pointDocument.population = 20; break;
-    }
-  }
-
-  /**
-   * Parse the first number out of a free-text OSM value like "4302", "14,115 ft", "1 000", "yes".
-   * Returns Double.NaN when there is no usable number. Never throws.
-   */
-  static double parseFirstNumber(String raw) {
-    if (raw == null) {
-      return Double.NaN;
-    }
-    StringBuilder sb = new StringBuilder();
-    boolean seenDigit = false;
-    boolean seenDot = false;
-    for (int i = 0; i < raw.length(); i++) {
-      char c = raw.charAt(i);
-      if (c >= '0' && c <= '9') {
-        sb.append(c);
-        seenDigit = true;
-      } else if (c == '.' && seenDigit && !seenDot) {
-        sb.append(c);
-        seenDot = true;
-      } else if ((c == ',' || c == ' ' || c == '\'') && seenDigit) {
-        // thousands separator within a number — skip it
-      } else if (seenDigit) {
-        break; // number ended (e.g. " ft", "-")
-      }
-    }
-    if (!seenDigit) {
-      return Double.NaN;
-    }
-    try {
-      return Double.parseDouble(sb.toString());
-    } catch (NumberFormatException e) {
-      return Double.NaN;
     }
   }
 
@@ -925,7 +291,8 @@ public class PlanetSearchProfile implements Profile {
       return null;
     }
     var pointDocument = new PointDocument();
-    setIconColorCategory(pointDocument, relation);
+    var relationTags = new OsmTagUtils.OsmTagCache(relation::getString);
+    setIconColorCategory(pointDocument, relation, relationTags);
 
     if (!"icon-river".equals(pointDocument.poiIcon) &&
         !"Bicycle".equals(pointDocument.poiCategory) &&
@@ -953,7 +320,7 @@ public class PlanetSearchProfile implements Profile {
     }
     var info = new RelationInfo(relation.id());
 
-    convertTagsToDocument(pointDocument, relation);
+    convertTagsToDocument(pointDocument, relation, relationTags);
 
     pointDocument.poiSource = "OSM";
     info.pointDocument = pointDocument;
@@ -1251,8 +618,9 @@ public class PlanetSearchProfile implements Profile {
       for (var mergedFeature : highway.getMergedFeatures()) {
         var minIdFeature = mergedFeature.representingFeature;
         var pointDocument = new PointDocument();
-        setIconColorCategory(pointDocument, minIdFeature);
-        convertTagsToDocument(pointDocument, minIdFeature);
+        var minIdTags = new OsmTagUtils.OsmTagCache(minIdFeature::getString);
+        setIconColorCategory(pointDocument, minIdFeature, minIdTags);
+        convertTagsToDocument(pointDocument, minIdFeature, minIdTags);
         pointDocument.poiSource = "OSM";
         pointDocument.poiLength = mergedFeature.length;
 
@@ -1281,12 +649,13 @@ public class PlanetSearchProfile implements Profile {
         : GeoUtils.point(feature.worldGeometry().getCoordinate());
 
     var pointDocument = new PointDocument();
-    convertTagsToDocument(pointDocument, feature);
+    var featureTags = new OsmTagUtils.OsmTagCache(feature::getString);
+    convertTagsToDocument(pointDocument, feature, featureTags);
     pointDocument.poiSource = "OSM";
     var lngLatPoint = GeoUtils.worldToLatLonCoords(point).getCoordinate();
     pointDocument.location = new double[] { lngLatPoint.getX(), lngLatPoint.getY() };
 
-    setIconColorCategory(pointDocument, feature);
+    setIconColorCategory(pointDocument, feature, featureTags);
 
     if (pointDocument.poiIcon == "icon-search") {
       return false;
@@ -1368,8 +737,8 @@ public class PlanetSearchProfile implements Profile {
     if (feature.hasTag("description")) {
       CoalesceIntoMap(pointDocument.description, "default", feature.getString("description"));
     }
-    // ADR-0011 item 1 — mirror of convertTagsToDocument's variant-name population for the
-    // non-icon (building/station/forest/...) path, which builds its own PointDocument inline.
+    // Mirror convertTagsToDocument's variant-name population for the non-icon path, which builds
+    // its own PointDocument inline.
     addAltNames(pointDocument, feature);
     pointDocument.wikidata = feature.getString("wikidata");
     pointDocument.image = feature.getString("image");
@@ -1380,23 +749,17 @@ public class PlanetSearchProfile implements Profile {
         : GeoUtils.point(feature.worldGeometry().getCoordinate());
     var lngLatPoint = GeoUtils.worldToLatLonCoords(point).getCoordinate();
     pointDocument.location = new double[] { lngLatPoint.getX(), lngLatPoint.getY() };
-    // This path builds the document inline and (unlike convertTagsToDocument) never classified it,
-    // so building/station/forest/natural=valley docs went in with feature_class=null. Classify here
-    // so they carry the same class-match ranking signal as every other emit path.
-    setFeatureClass(pointDocument, feature);
+    // This inline path never ran classification, so classify here too so building/station/forest
+    // docs carry the same class-match ranking signal as every other emit path.
+    pointDocument.feature_class = OsmTagUtils.classifyFeatureClass(feature::getString);
     insertPointToElasticsearch(pointDocument, docId);
   }
 
   /**
    * Safety floor for the insert path: every emitted document must carry a prominence so the
-   * query-time field_value_factor multiply is consistent. A few emit paths (e.g. ski-lift ways,
-   * relation-completion) don't run convertTagsToDocument, which would leave prominence null ->
-   * field_value_factor missing:1.0 -> they'd unfairly beat a real, scored feature (whose prominence
-   * is <1). Floor a null to the minimum prominence; leave any real value UNCHANGED. Package-private
-   * static so it can be unit-tested directly (extracted, behavior-preserving).
-   *
-   * <p>Enforces ADR-0001 (build-time-vs-query-time prominence split): "every emitted doc must carry
-   * a prominence value" — see docs/adr/0001-build-time-vs-query-time-prominence-split.md.
+   * query-time field_value_factor multiply is consistent. Some emit paths don't run
+   * convertTagsToDocument, leaving prominence null -> missing:1.0 -> they'd unfairly beat a real,
+   * scored feature. Floor a null to the minimum; leave any real value unchanged.
    */
   static float flooredProminence(Float prominence) {
     return prominence == null ? (float) ProminenceCalculator.FLOOR : prominence;
@@ -1404,8 +767,8 @@ public class PlanetSearchProfile implements Profile {
 
   private void insertPointToElasticsearch(PointDocument pointDocument, String docId) {
     pointDocument.prominence = flooredProminence(pointDocument.prominence);
-    emittedCount.increment();
-    emittedPointsCount.increment();
+    stats.emittedCount.increment();
+    stats.emittedPointsCount.increment();
     bulkIngester.add(BulkOperation.of(op -> op
         .index(idx -> idx
             .index(this.pointsIndexName)
@@ -1433,138 +796,86 @@ public class PlanetSearchProfile implements Profile {
         CoalesceIntoMap(bbox.name, "default", feature.getString("name"));
       }
       String bboxDocId = sourceFeatureToDocumentId(feature);
-      emittedCount.increment();
+      stats.emittedCount.increment();
       bulkIngester.add(BulkOperation.of(op -> op
           .index(idx -> idx
               .index(this.bboxIndexName)
               .id(bboxDocId)
               .document(bbox))));
     } catch (Exception e) {
-      // Only geometry/serialization building can throw here now; the actual
-      // indexing is handled (and its errors counted) by the bulk listener.
+      // Only geometry/serialization building can throw here; indexing errors are counted by the
+      // bulk listener.
       LOGGER.warning(() -> "Failed to build bbox document for "
           + sourceFeatureToDocumentId(feature) + ": " + e.getMessage());
     }
   }
 
   /**
-   * Flush every buffered document to Elasticsearch and wait for the in-flight
-   * bulk requests to complete. Must be called after planetiler.run() finishes
-   * emitting features and BEFORE the alias swap / refresh, so the new index is
-   * fully populated when it goes live.
-   *
-   * close() already flushes and waits internally, but we flush explicitly first
-   * (and log the remaining buffer) so the end-of-indexation flush is visible and
-   * intentional rather than an implicit side effect of close().
+   * Flush every buffered document and wait for in-flight bulk requests to complete. Call after
+   * planetiler.run() and before the alias swap / refresh, so the new index is fully populated when
+   * it goes live. close() also flushes, but we flush explicitly first so the final flush is visible.
    */
   public void flush() {
     LOGGER.info(() -> "Flushing final batch: " + bulkIngester.pendingOperations()
         + " buffered operations, " + bulkIngester.pendingRequests() + " in-flight requests.");
     bulkIngester.flush();
-    // close() blocks until the flushed buffer and all in-flight requests (and
-    // their listener callbacks, which update the counters) have completed.
+    // close() blocks until the buffer and all in-flight requests (and their counter-updating
+    // listener callbacks) have completed.
     bulkIngester.close();
-    LOGGER.info(() -> "Elasticsearch indexing finished: " + indexedCount.sum()
-        + " documents indexed, " + getFailedCount() + " failed ("
-        + failedPointsCount.sum() + " points, " + failedBboxCount.sum() + " bbox); "
-        + getTransientCharges() + " transient charge(s) after retries exhausted ("
-        + transientPointsCharges.sum() + " points, " + transientBboxCharges.sum() + " bbox).");
+    LOGGER.info(() -> "Elasticsearch indexing finished: " + stats.getIndexedCount()
+        + " documents indexed, " + stats.getFailedCount() + " failed ("
+        + stats.getFailedPointsCount() + " points, " + stats.getFailedBboxCount() + " bbox); "
+        + stats.getTransientCharges() + " transient charge(s) after retries exhausted ("
+        + stats.getTransientPointsCharges() + " points, " + stats.getTransientBboxCharges() + " bbox).");
   }
+
+  // Indexing accounting is owned by IndexingStats; these accessors delegate so MainClass (and tests)
+  // keep a stable surface on the profile.
 
   public long getIndexedCount() {
-    return indexedCount.sum();
+    return stats.getIndexedCount();
   }
 
-  /** Total failed documents across all indices (points + bbox). Keeps the
-   *  lossless invariant emitted == indexed + failed. */
   public long getFailedCount() {
-    return failedPointsCount.sum() + failedBboxCount.sum();
+    return stats.getFailedCount();
   }
 
-  /** Failed documents destined for the points index — these are missing SEARCH
-   *  results and must break the build. */
   public long getFailedPointsCount() {
-    return failedPointsCount.sum();
+    return stats.getFailedPointsCount();
   }
 
-  /** Failed bbox documents (geo_shape rejects on degenerate OSM geometries) —
-   *  warn-only; they don't affect name search. */
   public long getFailedBboxCount() {
-    return failedBboxCount.sum();
+    return stats.getFailedBboxCount();
   }
 
-  /** Points ops charged to the TRANSIENT bucket after retries were exhausted
-   *  (client-side connection/socket timeouts, 429/5xx). Likely phantom; tolerated
-   *  generously and backstopped by the reconcile gate. */
   public long getTransientPointsCharges() {
-    return transientPointsCharges.sum();
+    return stats.getTransientPointsCharges();
   }
 
-  /** Bbox ops charged to the transient bucket (warn-only analogue). */
   public long getTransientBboxCharges() {
-    return transientBboxCharges.sum();
+    return stats.getTransientBboxCharges();
   }
 
-  /** Total transient charges across both indices (points + bbox). */
   public long getTransientCharges() {
-    return transientPointsCharges.sum() + transientBboxCharges.sum();
+    return stats.getTransientCharges();
   }
 
   public long getEmittedCount() {
-    return emittedCount.sum();
+    return stats.getEmittedCount();
   }
 
-  /** Emitted POINTS documents only (excludes bbox). Denominator for the points
-   *  failure thresholds and the reconcile gate. */
   public long getEmittedPointsCount() {
-    return emittedPointsCount.sum();
+    return stats.getEmittedPointsCount();
   }
 
-  // ---------------------------------------------------------------------------
-  // Step D — two-bucket indexing guard.
-  //
-  // Replaces the all-or-nothing `failedPointsCount > 0` predicate that, during the
-  // incident, tripped exit-1 on 161 phantom transient timeouts even though the
-  // whole-planet index had been restored. Two independent thresholds now apply to
-  // POINTS:
-  //   * GENUINE per-item data failures (4xx mapping/parse) — these really are
-  //     missing search results, so the bar is strict: fail above
-  //     max(50, 0.0001% of emitted).
-  //   * TRANSIENT whole-batch charges (after retries exhausted) — likely phantom,
-  //     so tolerate up to max(5000, 0.05% of emitted); fail only above that.
-  // A real mass break (e.g. a mapping error dropping a big fraction) still fails
-  // loudly via the genuine-failure bucket. The reconcile gate is the final backstop.
-  // ---------------------------------------------------------------------------
-
-  /** Strict threshold for GENUINE per-item points data failures. */
-  static long genuineFailureThreshold(long emitted) {
-    return Math.max(50L, (long) Math.ceil(emitted * 0.000001)); // 0.0001%
-  }
-
-  /** Generous threshold for TRANSIENT points charges (likely-phantom timeouts). */
-  static long transientChargeThreshold(long emitted) {
-    return Math.max(5_000L, (long) Math.ceil(emitted * 0.0005)); // 0.05%
-  }
-
-  /**
-   * True when this run lost too many POINTS documents to treat the index as
-   * complete. Two buckets with independent thresholds (see above): a strict bar on
-   * genuine per-item data failures and a far more tolerant bar on transient
-   * whole-batch charges. bbox failures (geo_shape rejects) are warn-only and never
-   * trip this. Kept as a small pure predicate so the fail-the-build decision is
-   * unit-testable without calling {@code System.exit}.
-   */
+  /** Two-bucket points guard: see {@link IndexingStats#hasIndexingFailures()}. */
   public boolean hasIndexingFailures() {
-    long emittedPoints = emittedPointsCount.sum();
-    return failedPointsCount.sum() > genuineFailureThreshold(emittedPoints)
-        || transientPointsCharges.sum() > transientChargeThreshold(emittedPoints);
+    return stats.hasIndexingFailures();
   }
 
   /**
-   * Get the first point of the trail relation by checking some heuristics related
-   * to the relation's first member
-   * 
-   * @param mergedLines - the merged lines helper
+   * Get the first point of the trail relation via heuristics on its first member.
+   *
    * @return the first point of the trail relation
    * @throws GeometryException
    */
@@ -1593,17 +904,10 @@ public class PlanetSearchProfile implements Profile {
   }
 
   /**
-   * This method removes relation members that are part of super relations and
-   * have completed the ways processing.
-   * This is done by checking for each new way that is being processed if it
-   * completes a relation,
-   * and remove that relation from the list of parent relations, this way at some
-   * point all the ways and relations are empty
-   * and it means we can continue processing them to add them to the database and
-   * tiles.
-   * It also keeps track of the first and second member features in case they are
-   * needed to determine the first point of the relation.
-   * 
+   * Remove completed relation members from their parent super relations: when a new way completes a
+   * relation, drop that relation from its parents so that eventually all ways/relations are empty
+   * and can be emitted. Also tracks the first/second member features for finding the first point.
+   *
    * @param feature
    */
   private void handleSuperRelationMembersUpdate(SourceFeature feature) {
@@ -1712,15 +1016,23 @@ public class PlanetSearchProfile implements Profile {
         + feature.id();
   }
 
+  // Convenience overload for callers that don't already hold a per-feature tag cache (solo icon
+  // emits). Paths that also classify the same feature build ONE OsmTagCache and pass it to both this
+  // method and classifyFeatureClass, so the ~21 tag reads happen once per feature, not per method.
   private void setIconColorCategory(PointDocument pointDocument, WithTags feature) {
-    if ("protected_area".equals(feature.getString("boundary")) ||
-        "national_park".equals(feature.getString("boundary")) ||
-        "nature_reserve".equals(feature.getString("leisure"))) {
+    setIconColorCategory(pointDocument, feature, new OsmTagUtils.OsmTagCache(feature::getString));
+  }
+
+  private void setIconColorCategory(PointDocument pointDocument, WithTags feature,
+      OsmTagUtils.OsmTagCache tags) {
+    if ("protected_area".equals(tags.boundary) ||
+        "national_park".equals(tags.boundary) ||
+        "nature_reserve".equals(tags.leisure)) {
       setProtectedAreaIcon(pointDocument, feature);
       return;
     }
-    if (feature.getString("route") != null) {
-      switch (feature.getString("route")) {
+    if (tags.route != null) {
+      switch (tags.route) {
         case "hiking":
         case "foot":
           pointDocument.poiIconColor = "black";
@@ -1734,7 +1046,7 @@ public class PlanetSearchProfile implements Profile {
           pointDocument.poiCategory = "Bicycle";
           return;
         case "road":
-          if ("yes".equals(feature.getString("scenic"))) {
+          if ("yes".equals(tags.scenic)) {
             pointDocument.poiIconColor = "black";
             pointDocument.poiCategory = "4x4";
             pointDocument.poiIcon = "icon-four-by-four";
@@ -1742,10 +1054,10 @@ public class PlanetSearchProfile implements Profile {
           }
       }
     }
-    if (feature.getString("historic") != null) {
+    if (tags.historic != null) {
       pointDocument.poiIconColor = "#666666";
       pointDocument.poiCategory = "Historic";
-      switch (feature.getString("historic")) {
+      switch (tags.historic) {
         case "ruins":
           pointDocument.poiIcon = "icon-ruins";
           return;
@@ -1763,17 +1075,17 @@ public class PlanetSearchProfile implements Profile {
           return;
       }
     }
-    if ("picnic_table".equals(feature.getString("leisure")) ||
-        "picnic_site".equals(feature.getString("tourism")) ||
-        "picnic".equals(feature.getString("amenity"))) {
+    if ("picnic_table".equals(tags.leisure) ||
+        "picnic_site".equals(tags.tourism) ||
+        "picnic".equals(tags.amenity)) {
       pointDocument.poiIconColor = "#734a08";
       pointDocument.poiIcon = "icon-picnic";
       pointDocument.poiCategory = "Camping";
       return;
     }
 
-    if (feature.getString("natural") != null) {
-      switch (feature.getString("natural")) {
+    if (tags.natural != null) {
+      switch (tags.natural) {
         case "cave_entrance":
           pointDocument.poiIconColor = "black";
           pointDocument.poiIcon = "icon-cave";
@@ -1802,20 +1114,20 @@ public class PlanetSearchProfile implements Profile {
       }
     }
 
-    if ("reservoir".equals(feature.getString("water")) ||
-        "pond".equals(feature.getString("water")) ||
-        "lake".equals(feature.getString("water")) ||
-        "stream_pool".equals(feature.getString("water"))) {
+    if ("reservoir".equals(tags.water) ||
+        "pond".equals(tags.water) ||
+        "lake".equals(tags.water) ||
+        "stream_pool".equals(tags.water)) {
       pointDocument.poiIconColor = "#1e80e3";
       pointDocument.poiIcon = "icon-tint";
       pointDocument.poiCategory = "Water";
       return;
     }
 
-    if (feature.getString("man_made") != null) {
+    if (tags.manMade != null) {
       pointDocument.poiIconColor = "#1e80e3";
       pointDocument.poiCategory = "Water";
-      switch (feature.getString("man_made")) {
+      switch (tags.manMade) {
         case "water_well":
           pointDocument.poiIcon = "icon-water-well";
           return;
@@ -1825,29 +1137,29 @@ public class PlanetSearchProfile implements Profile {
       }
     }
 
-    if ("waterfall".equals(feature.getString("waterway"))) {
+    if ("waterfall".equals(tags.waterway)) {
       pointDocument.poiIconColor = "#1e80e3";
       pointDocument.poiIcon = "icon-waterfall";
       pointDocument.poiCategory = "Water";
       return;
     }
 
-    if ("waterway".equals(feature.getString("type"))) {
+    if ("waterway".equals(tags.type)) {
       pointDocument.poiIconColor = "#1e80e3";
       pointDocument.poiIcon = "icon-river";
       pointDocument.poiCategory = "Water";
       return;
     }
 
-    if (feature.getString("place") != null) {
+    if (tags.place != null) {
       pointDocument.poiIconColor = "black";
       pointDocument.poiIcon = "icon-home";
       pointDocument.poiCategory = "Wikipedia";
       return;
     }
 
-    if (feature.getString("tourism") != null) {
-      switch (feature.getString("tourism")) {
+    if (tags.tourism != null) {
+      switch (tags.tourism) {
         case "viewpoint":
           pointDocument.poiIconColor = "#008000";
           pointDocument.poiIcon = "icon-viewpoint";
@@ -1876,21 +1188,18 @@ public class PlanetSearchProfile implements Profile {
       }
     }
 
-    // Recall fix (issue #11): treat natural=volcano like natural=peak so named
-    // summit nodes tagged as volcanoes (e.g. Humphreys Peak, highest in AZ,
-    // OSM node 359267393) get icon-peak instead of falling through to the
-    // icon-search default, which causes processOtherSourceFeature to drop them
-    // (same missing-summit class as Mount Rainier). setFeatureClass and
-    // ProminenceCalculator already map volcano->peak; this completes that.
-    if ("peak".equals(feature.getString("natural")) || "volcano".equals(feature.getString("natural"))) {
+    // Treat natural=volcano like natural=peak so named summit nodes tagged as volcanoes get
+    // icon-peak instead of the icon-search default (which would let processOtherSourceFeature drop
+    // them). setFeatureClass and ProminenceCalculator already map volcano->peak.
+    if ("peak".equals(tags.natural) || "volcano".equals(tags.natural)) {
       pointDocument.poiIconColor = "black";
       pointDocument.poiIcon = "icon-peak";
       pointDocument.poiCategory = "Natural";
       return;
     }
 
-    if (feature.getString("highway") != null) {
-      switch (feature.getString("highway")) {
+    if (tags.highway != null) {
+      switch (tags.highway) {
         case "cycleway":
           pointDocument.poiIconColor = "black";
           pointDocument.poiCategory = "Bicycle";
@@ -1914,8 +1223,8 @@ public class PlanetSearchProfile implements Profile {
       }
     }
 
-    if ("place_of_worship".equals(feature.getString("amenity")) || "monastery".equals(feature.getString("amenity"))) {
-      var religion = feature.getString("religion") != null ? feature.getString("religion") : "";
+    if ("place_of_worship".equals(tags.amenity) || "monastery".equals(tags.amenity)) {
+      var religion = tags.religion != null ? tags.religion : "";
       pointDocument.poiCategory = "Other";
       pointDocument.poiIconColor = "black";
       switch (religion) {
@@ -1934,7 +1243,7 @@ public class PlanetSearchProfile implements Profile {
       }
     }
 
-    if (feature.getString("ref:IL:inature") != null) {
+    if (tags.refILInature != null) {
       pointDocument.poiIconColor = "#116C00";
       pointDocument.poiIcon = "icon-inature";
       pointDocument.poiCategory = "iNature";
@@ -1947,13 +1256,11 @@ public class PlanetSearchProfile implements Profile {
   }
 
   /**
-   * Differentiate protected areas by their PROTECTION TYPE so the search dropdown no longer shows the
-   * same green leaf for every national park, monument, wilderness and forest (the "grand …" case:
-   * Grand Canyon NP vs Grand Staircase-Escalante National Monument were both leaf). Keyed primarily
-   * off the free-text {@code protection_title} (the most reliable discriminator on US data), with
-   * {@code protect_class} and {@code boundary} as fallback. A plain national/state park or nature
-   * reserve keeps the leaf (the established meaning). Only icomoon glyphs that exist in the web font
-   * (src/fonts/icons.css) are used: icon-monument, icon-tree, icon-leaf.
+   * Differentiate protected areas by protection type so the dropdown no longer shows the same green
+   * leaf for every park, monument, wilderness and forest. Keyed primarily off the free-text
+   * protection_title (the most reliable discriminator on US data), with protect_class and boundary
+   * as fallback; a plain park or nature reserve keeps the leaf. Uses only glyphs that exist in the
+   * web font: icon-monument, icon-tree, icon-leaf.
    */
   private void setProtectedAreaIcon(PointDocument pointDocument, WithTags feature) {
     pointDocument.poiCategory = "Other";
