+++ src/test/java/il/org/osm/israelhiking/AddAltNamesTest.java
@@ -36,7 +36,7 @@ public class AddAltNamesTest {
     @Test
     public void splitsSemicolonSeparatedMultiValues() {
         // The single most likely bug: alt_name=A;B;C must become three searchable variants, not one.
-        var altNames = PlanetSearchProfile.buildAltNames(LANGS,
+        var altNames = OsmTagUtils.buildAltNames(LANGS,
                 tags(Map.of("alt_name", "Foo;Bar;Baz")));
         assertEquals(List.of("Foo", "Bar", "Baz"), altNames.get("default"),
                 "alt_name must be split on ';' and all variants kept");
@@ -44,7 +44,7 @@ public class AddAltNamesTest {
 
     @Test
     public void trimsWhitespaceAndDropsEmptyParts() {
-        var altNames = PlanetSearchProfile.buildAltNames(LANGS,
+        var altNames = OsmTagUtils.buildAltNames(LANGS,
                 tags(Map.of("alt_name", " Foo ;; Bar ;  ")));
         assertEquals(List.of("Foo", "Bar"), altNames.get("default"),
                 "parts must be trimmed and empty parts dropped");
@@ -55,7 +55,7 @@ public class AddAltNamesTest {
         var map = new HashMap<String, String>();
         map.put("alt_name", "Foo;Bar");
         map.put("official_name", "Bar;Qux"); // Bar repeats across tags
-        var altNames = PlanetSearchProfile.buildAltNames(LANGS, tags(map));
+        var altNames = OsmTagUtils.buildAltNames(LANGS, tags(map));
         assertEquals(List.of("Foo", "Bar", "Qux"), altNames.get("default"),
                 "duplicate variants across tags must be collapsed, insertion order preserved");
     }
@@ -63,7 +63,7 @@ public class AddAltNamesTest {
     @Test
     public void excludesOldName() {
         // old_name is deliberately NOT a source tag — a stale former name must never be searchable.
-        var altNames = PlanetSearchProfile.buildAltNames(LANGS,
+        var altNames = OsmTagUtils.buildAltNames(LANGS,
                 tags(Map.of("old_name", "Former Name")));
         assertNull(altNames, "old_name must be excluded entirely (no alt_names produced)");
     }
@@ -75,7 +75,7 @@ public class AddAltNamesTest {
         map.put("short_name", "Sh");
         map.put("loc_name", "Local");
         map.put("int_name", "International");
-        var altNames = PlanetSearchProfile.buildAltNames(LANGS, tags(map));
+        var altNames = OsmTagUtils.buildAltNames(LANGS, tags(map));
         var def = altNames.get("default");
         assertTrue(def.contains("Official"), "official_name must be indexed");
         assertTrue(def.contains("Sh"), "short_name must be indexed");
@@ -88,7 +88,7 @@ public class AddAltNamesTest {
         var map = new HashMap<String, String>();
         map.put("alt_name:he", "המיסדים;יונתן");
         map.put("alt_name:en", "IBT");
-        var altNames = PlanetSearchProfile.buildAltNames(LANGS, tags(map));
+        var altNames = OsmTagUtils.buildAltNames(LANGS, tags(map));
         assertEquals(List.of("המיסדים", "יונתן"), altNames.get("he"), "alt_name:he must key under 'he', split");
         assertEquals(List.of("IBT"), altNames.get("en"), "alt_name:en must key under 'en'");
         assertFalse(altNames.containsKey("default"), "no bare alt_name => no 'default' key");
@@ -97,7 +97,7 @@ public class AddAltNamesTest {
     @Test
     public void theIbtCase_altNameEnIbtIsSearchable() {
         // From the live OSM relation/19598237 (the client's IBT example): alt_name:en = IBT.
-        var altNames = PlanetSearchProfile.buildAltNames(LANGS,
+        var altNames = OsmTagUtils.buildAltNames(LANGS,
                 tags(Map.of("alt_name:en", "IBT")));
         assertEquals(List.of("IBT"), altNames.get("en"));
     }
@@ -105,7 +105,7 @@ public class AddAltNamesTest {
     @Test
     public void noVariantTagsYieldsNull() {
         // So addAltNames leaves alt_names null and @JsonInclude(NON_NULL) omits it from _source.
-        var altNames = PlanetSearchProfile.buildAltNames(LANGS,
+        var altNames = OsmTagUtils.buildAltNames(LANGS,
                 tags(Map.of("name", "Just A Name", "old_name", "stale")));
         assertNull(altNames, "a feature with no variant tags must produce no alt_names");
     }
